+++ crates/schematic/src/schema/renderers/toml_template.rs
@@ -96,7 +96,7 @@ impl<'gen> SchemaRenderer<'gen, String> for TomlTemplateRenderer<'gen> {
         false
     }
 
-    fn render_array(&mut self, array: &ArrayType) -> RenderResult<String> {
+    fn render_array(&mut self, array: &ArrayType, _schema: &Schema) -> RenderResult<String> {
         let key = self.ctx.get_stack_key();
 
         if !self.ctx.is_expanded(&key) {
@@ -108,31 +108,31 @@ impl<'gen> SchemaRenderer<'gen, String> for TomlTemplateRenderer<'gen> {
         Ok(format!("[{}]", self.render_schema(items_type)?))
     }
 
-    fn render_boolean(&mut self, boolean: &BooleanType) -> RenderResult<String> {
+    fn render_boolean(&mut self, boolean: &BooleanType, _schema: &Schema) -> RenderResult<String> {
         render_boolean(boolean)
     }
 
-    fn render_enum(&mut self, enu: &EnumType) -> RenderResult<String> {
+    fn render_enum(&mut self, enu: &EnumType, _schema: &Schema) -> RenderResult<String> {
         render_enum(enu)
     }
 
-    fn render_float(&mut self, float: &FloatType) -> RenderResult<String> {
+    fn render_float(&mut self, float: &FloatType, _schema: &Schema) -> RenderResult<String> {
         render_float(float)
     }
 
-    fn render_integer(&mut self, integer: &IntegerType) -> RenderResult<String> {
+    fn render_integer(&mut self, integer: &IntegerType, _schema: &Schema) -> RenderResult<String> {
         render_integer(integer)
     }
 
-    fn render_literal(&mut self, literal: &LiteralType) -> RenderResult<String> {
+    fn render_literal(&mut self, literal: &LiteralType, _schema: &Schema) -> RenderResult<String> {
         render_literal(literal)
     }
 
-    fn render_null(&mut self) -> RenderResult<String> {
+    fn render_null(&mut self, _schema: &Schema) -> RenderResult<String> {
         render_null()
     }
 
-    fn render_object(&mut self, object: &ObjectType) -> RenderResult<String> {
+    fn render_object(&mut self, object: &ObjectType, _schema: &Schema) -> RenderResult<String> {
         let key = self.ctx.get_stack_key();
         let value_type = self.ctx.resolve_schema(&object.value_type, &self.schemas);
 
@@ -157,7 +157,7 @@ impl<'gen> SchemaRenderer<'gen, String> for TomlTemplateRenderer<'gen> {
         Ok(format!("{{ {key} = {value} }}"))
     }
 
-    fn render_reference(&mut self, reference: &str) -> RenderResult<String> {
+    fn render_reference(&mut self, reference: &str, _schema: &Schema) -> RenderResult<String> {
         if let Some(schemas) = &self.schemas {
             if let Some(schema) = schemas.get(reference) {
                 return self.render_schema_without_reference(schema);
@@ -167,11 +167,11 @@ impl<'gen> SchemaRenderer<'gen, String> for TomlTemplateRenderer<'gen> {
         render_reference(reference)
     }
 
-    fn render_string(&mut self, string: &StringType) -> RenderResult<String> {
+    fn render_string(&mut self, string: &StringType, _schema: &Schema) -> RenderResult<String> {
         render_string(string)
     }
 
-    fn render_struct(&mut self, structure: &StructType) -> RenderResult<String> {
+    fn render_struct(&mut self, structure: &StructType, _schema: &Schema) -> RenderResult<String> {
         let mut out = vec![];
 
         for field in &structure.fields {
@@ -193,15 +193,15 @@ impl<'gen> SchemaRenderer<'gen, String> for TomlTemplateRenderer<'gen> {
         Ok(out.join(self.ctx.gap()))
     }
 
-    fn render_tuple(&mut self, tuple: &TupleType) -> RenderResult<String> {
+    fn render_tuple(&mut self, tuple: &TupleType, _schema: &Schema) -> RenderResult<String> {
         render_tuple(tuple, |schema| self.render_schema(schema))
     }
 
-    fn render_union(&mut self, uni: &UnionType) -> RenderResult<String> {
+    fn render_union(&mut self, uni: &UnionType, _schema: &Schema) -> RenderResult<String> {
         render_union(uni, |schema| self.render_schema(schema))
     }
 
-    fn render_unknown(&mut self) -> RenderResult<String> {
+    fn render_unknown(&mut self, _schema: &Schema) -> RenderResult<String> {
         render_unknown()
     }
 
@@ -213,6 +213,7 @@ impl<'gen> SchemaRenderer<'gen, String> for TomlTemplateRenderer<'gen> {
         self.schemas = Some(schemas);
 
         let mut root = validate_root(schemas)?;
+        let fake_schema = Schema::default();
 
         // Recursively extract all sections (arrays, objects)
         if let SchemaType::Struct(doc) = &mut root.type_of {
@@ -226,7 +227,7 @@ impl<'gen> SchemaRenderer<'gen, String> for TomlTemplateRenderer<'gen> {
             sections.push(format!(
                 "{}[[{key}]]\n{}",
                 value.comment,
-                self.render_struct(&value.table)?
+                self.render_struct(&value.table, &fake_schema)?
             ));
         }
 
@@ -234,7 +235,7 @@ impl<'gen> SchemaRenderer<'gen, String> for TomlTemplateRenderer<'gen> {
             sections.push(format!(
                 "{}[{key}]\n{}",
                 value.comment,
-                self.render_struct(&value.table)?
+                self.render_struct(&value.table, &fake_schema)?
             ));
         }